+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [4.54.0](https://github.com/plivo/plivo-php/tree/v7.33.0) (2023-08-07)
+**Feature - WhatsApp message support**
+- Added new param `template` and  new message_type `whatsapp` to [send message API](https://www.plivo.com/docs/sms/api/message#send-a-message)
+- Added  new  `message_states` (`read`)   `message_type`(`whatsapp`),`conversation_id`, `conversation_origin`, `conversation_expiry_timestamp` in [list all messages API](https://www.plivo.com/docs/sms/api/message#list-all-messages) and [get message details API](https://www.plivo.com/docs/sms/api/message#retrieve-a-message) response
+
 ## [4.53.0](https://github.com/plivo/plivo-php/tree/v4.53.0) (2023-08-03)
 **Feature - DLT parameters**
 - Added new params `DLTEntityID`, `DLTTemplateID`, `DLTTemplateCategory` to the [send message API](https://www.plivo.com/docs/sms/api/message/send-a-message/)